server/pyproject.toml
[tool.poetry]
name = "Digital_Wallet"
version = "0.3.1"
description = "digital wallet management application for a fintech platform that allows users to create and manage wallets."
authors = ["Siri <sirilakshmi1703@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
motor = "^3.6"
msgspec = "^0.18.6"
newrelic = "^9.7.0"
orjson = "^3.9.15"
pydantic = "^2.6.2"
pydantic-settings = "^2.2.1"
pymongo = "^4.6.2"
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
pyjwt = "^2.6.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
python-multipart = "^0.0.16"
